bin/switch-site script for easily switching configs
Given a set of config files of the format `conf/general-{cobrand}.yml`
this script lets you easily create and recreate a symlink at
`conf/general.yml` that points to the cobrand config of your choice.

I’ve been using this for years; never thought to commit it!
Thanks to @davea for the reminder.