Remove Bower and Grunt dependencies
Jekyll has come a long way since this branch was first set up. It can
now compile Sass files on its own, and `jekyll serve` now watches for
filesystem changes by default.

With Jekyll compiling the Sass by itself, installing npm and Grunt just
to concatenate some JavaScript seemed fairly overblown. So this commit
removes the Grunt dependency (and the JavaScript concatenation).

While we were at it, we also removed the Bower dependency. The site
only uses 3 JavaScript libraries, and given the rarity with which
the requirements are ever updated, it seemed a greater overhead to
expect contributors to install npm and Bower, than it is just to
include the third-party .js files in `/assets/scripts`.

We also provide a Gemfile (for Bundler), so contributors can ensure
their dev environment matches the production environment (Github Pages).